==> demo_website/01_loaddata.qmd <==
---
title: "01_loaddata"
editor: visual
editor_options: 
  chunk_output_type: console
---

# load packages, set directories

```{r}
#| warning: false
library(SingleCellExperiment)
library(DropletUtils)
library(scRNAseq)

rawdata_dir <- "data/raw"
outdata_dir <- "data/output"
```

# read in dataset

```{r}
sce <- read10xCounts(file.path(rawdata_dir, "LC026"))
```

# save RDS

```{r}
saveRDS(sce, file = file.path(outdata_dir, "sce_unfiltered.rds"))
```

# or, find a dataset in the `scRNAseq` collection

```{r}
ld <- listDatasets()

sce_cortex <- ZhongPrefrontalData()
```

# save RDS

```{r}
saveRDS(sce_cortex, file = file.path(outdata_dir, "sce_cortex.rds"))
```

# sessioninfo

```{r}
sessionInfo()
```
