==> demo_website/07_advanced-miloR-trajectory.qmd <==
---
title: "07_advanced-miloR-trajectory"
editor: visual
editor_options: 
  chunk_output_type: console
---

# load packages, set directories

```{r}
#| warning: false
library(SingleCellExperiment)
library(miloR)
library(dplyr)
library(patchwork)
library(scran)
library(scater)
library(bluster)
library(scDblFinder)
library(slingshot)
library(tradeSeq)


data_dir <- "data/output"
```

# read in dataset

```{r}
# already-manually-annotated dataset
sce <- readRDS(file.path(data_dir, "3-SCE_clustered.rds"))
```

# `miloR` analysis

```{r}

# https://bioconductor.org/packages/release/bioc/vignettes/miloR/inst/doc/milo_demo.html

sce <- Milo(sce)

sce <- buildGraph(sce, k = 20, d = 30)
sce <- makeNhoods(sce, k=20, d=30, prop=0.1, refined=TRUE)

plotNhoodSizeHist(sce) 

sce <- countCells(sce,
                  meta.data=data.frame(colData(sce)),
                  sample="sample_id")
sce <- calcNhoodDistance(sce, d=30)

(nhc <- nhoodCounts(sce))

# organize metadata
md <- data.frame(sample_id = colnames(nhc))
md$group <- sapply(strsplit(md$sample_id,"_"), .subset, 2)
rownames(md) <- colnames(nhc)
md


 
milo_res <- testNhoods(sce, design=~group, design.df=md)

milo_res %>% arrange(SpatialFDR) %>% select(-logCPM,-FDR)

sce <- buildNhoodGraph(sce)
plotUMAP(sce) + 
  plotNhoodGraphDA(sce, milo_res, alpha=0.05) +
  plot_layout(guides="collect")

```

# grab an appropriate example trajectory dataset

```{r}

# some code taken from 
# https://nbisweden.github.io/workshop-archive/workshop-scRNAseq/2020-01-27/labs/compiled/slingshot/slingshot.html

# url <- "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE72nnn/GSE72857/suppl/GSE72857_umitab.txt.gz"
# f <- file.path("data/output",basename(url))
# download.file(url, destfile = f)
# data <- read.delim(gzfile(f), header = TRUE, row.names = 1)
# object.size(data)
# 
# data <- Matrix::Matrix(as.matrix(data), sparse = TRUE)
# object.size(data)
# saveRDS(data, file=file.path(data_dir, "data-GSE72857.rds"))
# > 166318016/1135197256
# [1] 0.1465102

# data <- readRDS(file.path(data_dir, "data-GSE72857.rds"))
# sce <- SingleCellExperiment(assays = list(counts=data))
# sce <- logNormCounts(sce)


# in data/output/ dir
# wget https://github.com/hemberg-lab/nrg-paper-figures/raw/master/deng-reads.rds

sce <- readRDS(file.path(data_dir, "deng-reads.rds"))

mgv <- modelGeneVar(sce)
hvgs <- getTopHVGs(mgv, n = 1500)

sce <- runPCA(sce, subset_row = hvgs, ncomponents = 5)
system.time(sce <- runUMAP(sce, dimred = "PCA"))

plotUMAP(sce)
plotPCA(sce, colour_by = "cell_type1")

# clust1 <- clusterCells(sce, use.dimred="PCA",
#                       BLUSPARAM=NNGraphParam(cluster.fun="louvain",
#                                              k=10))

clust <- fastcluster(sce, rdname = "PCA", 
                     k = 10, returnType = "preclusters")

colData(sce)$kmeans_15 <- as.factor(clust)

# plotUMAP(sce, colour_by="clusters")
plotPCA(sce, colour_by="kmeans_15")



```

# `slingshot` analysis

```{r}

# https://bioconductor.org/packages/devel/bioc/vignettes/slingshot/inst/doc/vignette.html
sce <- slingshot(sce, reducedDim = 'PCA', 
                 clusterLabels = "kmeans_15")
sce

plotPCA(sce, colour_by = "slingPseudotime_1")



# fit negative binomial GAM
sce <- fitGAM(sce)

# test for dynamic expression
ATres <- associationTest(sce)



```

# sessioninfo

```{r}
sessionInfo()
```
